Removed Code on Pages to perform Collab Push
