+++ lib/oil.dart
@@ -1,125 +1,1 @@
-import 'package:flutter/material.dart';
-import 'package:recycle/scan.dart';
 
-final String mainFont = "AppleGothic";
-
-void main() {
-  runApp(const Oil());
-}
-
-//class
-class Oil extends StatelessWidget {
-  const Oil({super.key});
-  //color of background
-  static const color = Color(0xFFB6E8C6);
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'no title',
-      theme: ThemeData(
-        elevatedButtonTheme: ElevatedButtonThemeData(
-          style: ElevatedButton.styleFrom(foregroundColor: Colors.black),
-        ),
-        //sets the background color of the scene completely
-        scaffoldBackgroundColor: const Color(0xFFB6E8C6),
-        fontFamily: mainFont,
-      ),
-      //homepage text
-      home: const OilPage(title: 'Oil'),
-    );
-  }
-}
-
-class OilPage extends StatefulWidget {
-  const OilPage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<OilPage> createState() => _OilPageState();
-}
-
-class _OilPageState extends State<OilPage> {
-  /* This is the Area of the Project where you set up the Structure of the
-  app.
-  */
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      /*there is an app bar that acts as a divider but because we set up the
-     same color as the background we can can't tell the difference
-     as a test, hover over the hex code and use another color. 
-     */
-      body: Center(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            SizedBox(height: 20),
-            SizedBox(
-              width: 75.0,
-              height: 150.0,
-              /*if you are adding a component inside the sized box then
-              you must declare it as a child followed by closing comma etc
-              */
-              child: Image(image: AssetImage('assets/recycling.png')),
-            ),
-            SizedBox(),
-            RichText(
-              text: TextSpan(
-                text: 'This item contains Oil',
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 20,
-                  fontWeight: null,
-                  fontFamily: mainFont,
-                ),
-              ),
-            ),
-            SizedBox(height: 50),
-            SizedBox(
-              child: RichText(
-                text: TextSpan(
-                  text: "",
-                  style: TextStyle(
-                    color: Colors.black,
-                    fontSize: 15,
-                    fontWeight: null,
-                    fontFamily: mainFont,
-                  ),
-                ),
-              ),
-            ),
-            ElevatedButton(
-              style: const ButtonStyle(
-                backgroundColor: WidgetStatePropertyAll<Color>(
-                  Color.fromARGB(255, 169, 240, 191),
-                ),
-              ),
-              child: const Text('Return to Scan another Item'),
-              onPressed: () {
-                Navigator.of(context).push(_switchToScan());
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-Route _switchToScan() {
-  return PageRouteBuilder(
-    pageBuilder: (context, animation, secondaryAnimation) => const ScanPage(),
-    transitionsBuilder: (context, animation, secondaryAnimation, child) {
-      const begin = Offset(0.0, 1.0);
-      const end = Offset.zero;
-      const curve = Curves.ease;
-
-      var tween = Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
-
-      return SlideTransition(position: animation.drive(tween), child: child);
-    },
-  );
-}
